Added some unit tests for build_llvm. Added tests for kernel_source. Moved test for get_sources_with_params to tests for kernel_source.
